requirements.txt
langchain==0.3.18
langchain-core==0.3.36
langchain-community==0.3.17
langchain-groq==0.2.4
langchain-huggingface==0.1.2
langchain-chroma==0.2.2
langchain-text-splitters==0.3.6
chromadb>=0.6.3
huggingface-hub>=0.29.1
groq==0.13.1
pydantic>=2.10.6
flask==3.0.3
python-dotenv==1.0.1
presidio-analyzer==2.2.355
presidio-anonymizer==2.2.355
spacy==3.7.5
sentence-transformers==3.0.1
transformers==4.44.2
torch==2.4.1
pandas==2.2.2
numpy==1.26.4
PyYAML==6.0.2
requests==2.32.3
flask-cors==4.0.1
flask-login==0.6.3
werkzeug==3.0.3
jinja2==3.1.4
markupsafe==2.1.5